LPS-58771 Create a register to hold bundles that provides the soy capability
