+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/internal/SoyProviderCapabilityBundleRegister.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.template.soy.internal;
+
+import com.liferay.portal.template.soy.utils.SoyTemplateUtil;
+
+import java.util.Collection;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+import org.osgi.framework.Bundle;
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Rafael Praxedes
+ */
+@Component(service = SoyProviderCapabilityBundleRegister.class)
+public class SoyProviderCapabilityBundleRegister {
+
+	public Bundle getBundle(long bundleId) {
+		return _bundleMap.get(bundleId);
+	}
+
+	public Collection<Bundle> getBundles() {
+		return _bundleMap.values();
+	}
+
+	public Bundle getTemplateBundle(String templateId) {
+		long bundleId = SoyTemplateUtil.getBundleId(templateId);
+
+		Bundle bundle = getBundle(bundleId);
+
+		if (bundle == null) {
+			throw new IllegalStateException(
+				"There are no bundles providing " + templateId);
+		}
+
+		return bundle;
+	}
+
+	public void register(Bundle bundle) {
+		_bundleMap.put(bundle.getBundleId(), bundle);
+	}
+
+	public void unregister(Bundle bundle) {
+		_bundleMap.remove(bundle.getBundleId());
+	}
+
+	private static final Map<Long, Bundle> _bundleMap =
+		new ConcurrentHashMap<>();
+
+}
